third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEOffsetElement-dom-in-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' attribute of the SVGFEOffsetElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS offsetElement.getAttribute('in') is "blur"
PASS offsetElement.getAttribute('in') is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
